-Fix a problem casting a pointer to an int in the master server under 64-bit Windows. Thanks to Cray Elliot for the fix (bug 978).
SVN r5008 (trunk)

+++ master/main.cpp
@@ -365,7 +365,7 @@ void pingServer(SServer &s)
 	}
 
 #ifdef _WIN32
-	s.key_sent = rand() * (int)GetModuleHandle(0) * time(0);
+	s.key_sent = rand() * (intptr_t)GetModuleHandle(0) * time(0);
 #else
 	s.key_sent = rand() * getpid() * time(0);
 #endif
